+++ src/Hangfire.Mongo/Migration/Strategies/Backup/NoneMongoBackupStrategy.cs
@@ -1,0 +1,11 @@
+﻿using MongoDB.Driver;
+
+namespace Hangfire.Mongo.Migration.Strategies.Backup
+{
+    /// <summary>
+    /// No backup strategy
+    /// </summary>
+    public class NoneMongoBackupStrategy : MongoBackupStrategy
+    {
+    }
+}